# src/main/resources/OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/article/{id}": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "indexName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "controller"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "indexName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/article": {
      "put": {
        "tags": [
          "controller"
        ],
        "operationId": "addArticle",
        "parameters": [
          {
            "name": "indexName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "controller"
        ],
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "indexName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/indexStatus": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "checkIndexStatus",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndexStatus"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clearIndex": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "clearIndex",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/initTrec": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "initTrecData",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "getArticles",
        "parameters": [
          {
            "name": "indexName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/initData": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "initData",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/savedIndex": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "checkSavedIndexState",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedIndexStatus"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/loadIndex": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "loadIndex",
        "parameters": [
          {
            "name": "fileName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/saveIndex": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "saveIndex",
        "parameters": [
          {
            "name": "fileName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/query": {
      "post": {
        "tags": [
          "controller"
        ],
        "operationId": "search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArticleModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "type": "string"
          },
          "downloadDate": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "published": {
            "type": "string",
            "format": "date-time"
          },
          "rank": {
            "type": "integer",
            "format": "int32"
          },
          "score": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "QueryResultModel": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleModel"
            }
          },
          "searchWords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "documentsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "IndexStatus": {
        "type": "object",
        "properties": {
          "initialized": {
            "type": "boolean"
          },
          "indexName": {
            "type": "string"
          }
        }
      },
      "SavedIndexStatus": {
        "type": "object",
        "properties": {
          "saved": {
            "type": "boolean"
          },
          "indexName": {
            "type": "string"
          }
        }
      },
      "QueryModel": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "vectorModel": {
            "type": "boolean"
          },
          "numberOfResults": {
            "type": "integer",
            "format": "int32"
          },
          "indexName": {
            "type": "string"
          }
        }
      }
    }
  }
}